Made changes to rcn_detail.html, and other pages
I have set us the rcn_detail page so that all we have to do is put in
the necessary back end links, and in the logged in pages I have carried
on the basic style of the log in page. All the pages now look a little
better. I need you to make some changes so that the page will grab
information from the back end, such as the times, and the crew number
which can then be embedded into the id of the button.