Iphone compatability started
If you come to the website using an iphone (and therefore sporting the
iphone agent) you will be switched to a special new CSS page which will
make the website more touchscreen friendly. I have not made the website
fully work yet, but it is definitely a good start. i need to start
testing it on an actual Iphone screen.