aced the iPad, and mostly the iPhone layout
It’s going really well, the only problem is that there is some stuff to
be fixed on the PC view still, and I have introduced a button now but
it doesn’t work yet (the detail button on the RCN main page). Also for
some reason the get times button has stopped working, something needs
to be done for that. But otherwise the IOS version is going very well.